% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacitance.R
\name{capacitance}
\alias{capacitance}
\alias{lung_capacitance}
\title{Capacitance}
\usage{
capacitance(
  tissue_volume,
  tissue_coefficient,
  blood_volume,
  blood_coefficient,
  tp_factor = stp_factor()
)

lung_capacitance(
  air_volume,
  tissue_volume,
  tissue_coefficient,
  blood_volume,
  blood_coefficient,
  tp_factor = stp_factor()
)
}
\arguments{
\item{tissue_volume}{\code{numeric}, tissue volume in l.}

\item{tissue_coefficient}{\code{numeric}, tissue:gas partition coefficient}

\item{blood_volume}{\code{numeric}, tissue volume in l.}

\item{blood_coefficient}{\code{numeric}, blood:gas partition coefficient}

\item{tp_factor}{\code{numeric}, temperature/pressure factor}

\item{air_volume}{\code{numeric}, air volume in l.}
}
\value{
\code{numeric}, conductance of the tissue in l/(min*atm)
}
\description{
Capacitance is defined as ability of a tissue to hold anaesthetic agents
ant any given partial pressure.
}
\details{
From Cowles et al.:
"KETY(14) has shown that most tissues are so finely perfused that they
can be considered to be in equilibrium with their venous blood.
Therefore the venous blood contained within a tissue compartment must be
considered to be a part of that tissue compartment.
}
\references{
\code{capacitance}: Eqn. 14 in
Cowles, A. L., Borgstedt, H. H., & Gillies, A. J. (1973).
A simplified digital method for predicting anesthetic uptake and distribution.
Computers in Biology and Medicine, 3(4), 385-395.
\doi{10.1016/0010-4825(73)90004-8}

\code{lung_capacitance}: Eqn. 15 in
Cowles, A. L., Borgstedt, H. H., & Gillies, A. J. (1973).
A simplified digital method for predicting anesthetic uptake and distribution.
Computers in Biology and Medicine, 3(4), 385-395.
\doi{10.1016/0010-4825(73)90004-8}
}
