% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simva.R
\name{cardiac_output}
\alias{cardiac_output}
\title{Cardiac output}
\usage{
cardiac_output(
  total = 6.3,
  prop_lung = 1,
  prop_vrg = 0.798,
  prop_mus = 0.157,
  prop_fat = 0.044
)
}
\arguments{
\item{total}{\code{numeric}, total cardiac output in l/min.}

\item{prop_lung}{\code{numeric}, proportion of cardiac output to the lung, should
be 1.00 in general.}

\item{prop_vrg}{\code{numeric}, proportion of cardiac output to the vessel rich
compartment.}

\item{prop_mus}{\code{numeric}, proportion of cardiac output to the musle
compartment.}

\item{prop_fat}{\code{numeric}, proportion of cardiac output to the fat
compartment.}
}
\value{
a named \code{numeric} with blood flow values (in l/min) for
\emph{lung}, \emph{vrg} (vessel rich group, visceral: brain, heart, kidney, ...),
\emph{mus} (muscle, lean tissue: muscle, skin, subcutaneous tissue, ...),
and \emph{fat} (fat, fatty tissue: yellow marrow, fat) for the given anaesthetic.
}
\description{
Partial blood flow of common compartments.
}
\references{
Table 3 in
Cowles, A. L., Borgstedt, H. H., & Gillies, A. J. (1973).
A simplified digital method for predicting anesthetic uptake and distribution.
Computers in Biology and Medicine, 3(4), 385-395.
\doi{10.1016/0010-4825(73)90004-8}
}
