% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simva.R
\name{lung_capacitance}
\alias{lung_capacitance}
\title{Lung capacitance}
\usage{
lung_capacitance(
  air_volume,
  tissue_volume,
  tissue_coefficient,
  blood_volume,
  blood_coefficient,
  tp_factor = stp_factor()
)
}
\arguments{
\item{air_volume}{\code{numeric}, air volume in l.}

\item{tissue_volume}{\code{numeric}, tissue volume in l.}

\item{tissue_coefficient}{\code{numeric}, tissue:gas partition coefficient}

\item{blood_volume}{\code{numeric}, tissue volume in l.}

\item{blood_coefficient}{\code{numeric}, blood:gas partition coefficient}

\item{tp_factor}{\code{numeric}, temperature/pressure factor}
}
\value{
\code{numeric}, conductance of the tissue in l/(min*atm)
}
\description{
Capacitance is defined as ability of a tissue to hold anaesthetic agents
ant any given partial pressure.
}
\details{
From Cowles et al.:
Alveolar air, lung tissue and arterial blood are also in equilibrium.
}
\references{
Eqn. 15 in
Cowles, A. L., Borgstedt, H. H., & Gillies, A. J. (1973).
A simplified digital method for predicting anesthetic uptake and distribution.
Computers in Biology and Medicine, 3(4), 385-395.
\doi{10.1016/0010-4825(73)90004-8}
}
