% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{partial_pressures}
\alias{partial_pressures}
\title{Partial Pressures}
\usage{
partial_pressures(pinsp, lung = 0, vrg = 0, mus = 0, fat = 0, cv = 0)
}
\arguments{
\item{pinsp}{\code{double(1)}, inspired anaesthetic partial pressure in atm.}

\item{lung}{\code{double(1)}, partial pressure in the lung.}

\item{vrg}{\code{double(1)}, partial pressure in the vessel rich group
compartment.}

\item{mus}{\code{double(1)}, partial pressure in the lean/muscular
compartment.}

\item{fat}{\code{double(1)}, partial pressure in the fat
compartment.}

\item{cv}{\code{double(1)}, partial pressure in the central venous
compartment.}
}
\value{
a named \code{numeric} with partial pressures suiteable as input for the
\code{\link[=sim_anaesthetic_uptake]{sim_anaesthetic_uptake()}} function.
}
\description{
Create initial vector of partial pressures for
\code{\link[=sim_anaesthetic_uptake]{sim_anaesthetic_uptake()}}.
}
\examples{
partial_pressures(pinsp = 12)
}
