% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{simva}
\alias{simva-package}
\alias{simva}
\title{Simulate Volatile Anaesthetic Uptake and Distribution}
\description{
Implementation of the simple three compartment model for volatile
anaesthetics described in Cowles et al. 1973.
}
\details{
The main function is \code{\link[=sim_anaesthetic_uptake]{sim_anaesthetic_uptake()}}.
}
\references{
Eqn. 6 in
Cowles, A. L., Borgstedt, H. H., & Gillies, A. J. (1973).
A simplified digital method for predicting anesthetic uptake and distribution.
Computers in Biology and Medicine, 3(4), 385-395.
\doi{10.1016/0010-4825(73)90004-8}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/umg-minai/simva/}
  \item Report bugs at \url{https://github.com/umg-minai/simva/issues/}
}

}
\author{
\strong{Maintainer}: Sebastian Gibb \email{mail@sebastiangibb.de} (\href{https://orcid.org/0000-0001-7406-4443}{ORCID})

}
\keyword{internal}
