% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{stp_factor}
\alias{stp_factor}
\title{Standard temperature and pressure correction factor}
\usage{
stp_factor(
  std_temperature = 273.2,
  tissue_temperature = 310.2,
  std_pressure = 1
)
}
\arguments{
\item{std_temperature}{\code{numeric}, standard temperature value in Kelvin,
default: 273.2 K (== 0 °C)}

\item{tissue_temperature}{\code{numeric}, tissue temperature value in Kelvin,
default: 310.2 (== 37 °C)}

\item{std_pressure}{\code{numeric}, standard_pressure in atm, default: 1 atm
(== 101.325 kPa)}
}
\value{
\code{numeric}, correction factor for temperature/pressure condition
}
\description{
Standard temperature and pressure correction factor
}
\examples{
# BTPS (body temperature, pressure saturated (ignored))
stp_factor()
# STPD (standard temperature, pressure, dry)
stp_factor(tissue_temperature = 273.2)
}
