% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simva.R
\name{tissue_coefficients}
\alias{tissue_coefficients}
\title{Partition coefficients}
\usage{
tissue_coefficients(anaesthetic = c("nitrous-oxide", "diethyl-ether"))
}
\arguments{
\item{anaesthetic}{\code{character(1)}, anaesthetic}
}
\value{
a named \code{numeric} with tissue partition coefficients for
\emph{lung}, \emph{vrg} (vessel rich group, visceral: brain, heart, kidney, ...),
\emph{mus} (muscle, lean tissue: muscle, skin, subcutaneous tissue, ...),
and \emph{fat} (fat, fatty tissue: yellow marrow, fat) for the given anaesthetic.
}
\description{
Partition coefficients for commonly used volatile anaesthetics.
}
\examples{
tissue_coefficients("nitrous-oxide")
}
\references{
Table 2 in
Cowles, A. L., Borgstedt, H. H., & Gillies, A. J. (1973).
A simplified digital method for predicting anesthetic uptake and distribution.
Computers in Biology and Medicine, 3(4), 385-395.
\doi{10.1016/0010-4825(73)90004-8}
}
