% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{partial_pressures}
\alias{partial_pressures}
\title{Partial Pressures}
\usage{
partial_pressures(
  pinsp,
  palv = 0,
  part = 0,
  pvrg = 0,
  pmus = 0,
  pfat = 0,
  pcv = 0
)
}
\arguments{
\item{pinsp}{\code{double(1)}, inspired anaesthetic partial pressure in atm.}

\item{palv}{\code{double(1)}, alveolar pressure, partial pressure in the lung.}

\item{part}{\code{double(1)}, arterial partial pressure.}

\item{pvrg}{\code{double(1)}, partial pressure in the vessel rich group
compartment.}

\item{pmus}{\code{double(1)}, partial pressure in the lean/muscular
compartment.}

\item{pfat}{\code{double(1)}, partial pressure in the fat
compartment.}

\item{pcv}{\code{double(1)}, partial pressure in the central venous
compartment.}
}
\value{
a named \code{numeric} with partial pressures suiteable as input for the
\code{\link[=sim_anaesthetic_uptake]{sim_anaesthetic_uptake()}} function.
}
\description{
Create initial vector of partial pressures for
\code{\link[=sim_anaesthetic_uptake]{sim_anaesthetic_uptake()}}.
}
\examples{
partial_pressures(pinsp = 12)
}
